+++ NEWS.md
@@ -1,9 +1,10 @@
-# simmer.plot 0.1.17.3
+# simmer.plot 0.1.17.4
 
 - Fix rollback parsing for simmer > 4.4.5.
 - Fix test warnings due to deprecated features.
 - Remove deprecated methods `plot.simmer` and `plot.wrap`.
 - Fix clone postprocessing when `n` is specified as a function (#27).
+- Add support for activity tags and named rollbacks (#26).
 
 # simmer.plot 0.1.17